Test Maven project, manually make a jar-file in the maven project and creating a manifest.txt
The Maven project was tested, if everything works while creating a executable jar.
In the second step, a executable jar-file was created manually. To do this we also add a "manifest.txt".
In the end the .gitignore-file was changed, to not push the wrong files and directories in the
repository.